+++ authentication/api/exceptions.py
@@ -6,24 +6,6 @@ class KeyNotFoundError(FrameworkAuthError):
     pass
 
 
-class CertificateError(FrameworkAuthError):
-    """
-    Base class for errors related to the client certificate
-    """
-
-
-class CertificateMissingError(CertificateError):
-    pass
-
-
-class CertificateRoleError(CertificateError):
-    pass
-
-
-class CertificateRoleMissingError(CertificateRoleError):
-    pass
-
-
 class AccessTokenDecodingError(FrameworkAuthError):
     """
     Base class for errors related to the issuing of the token